backend/pyproject.toml
[project]
name="obc-api"
version="1.0"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "bcrypt==3.2.0",
    "beautifulsoup4==4.14.2",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==46.0.2",
    "dnspython==2.8.0",
    "ecdsa==0.19.1",
    "email-validator==2.3.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.11",
    "fastapi-cloud-cli==0.1.5",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mypy-extensions==1.1.0",
    "nodeenv==1.9.1",
    "packaging==25.0",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "platformdirs==4.4.0",
    "psycopg==3.2.10",
    "psycopg-binary==3.2.10",
    "psycopg-pool==3.2.6",
    "pyasn1==0.6.1",
    "pycparser==2.23",
    "pydantic==2.11.9",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "requests==2.32.5",
    "rich==14.1.0",
    "rich-toolkit==0.15.1",
    "rignore==0.6.4",
    "rsa==4.9.1",
    "sentry-sdk==2.37.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "starlette==0.47.3",
    "tomlkit==0.13.3",
    "typer==0.17.4",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
]

[tool.isort]
profile = "black"
